+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Improvement:
 - [#1089](https://github.com/elysiajs/elysia/pull/1089) add stop function to ElysiaAdapter type
 
 Bug fix:
-- [#1126](https://github.com/elysiajs/elysia/pull/1126) Websocket errors not catching
+- [#1126](https://github.com/elysiajs/elysia/pull/1126) websocket errors not catching
+- [#1281](https://github.com/elysiajs/elysia/issues/1281) automatically enforce additional properties in nested schema (eg. array)
 
 # 1.3.5 - 16 Jun 2025
 Bug fix: